Add simple drive to point controller test.
