+++ src/amberdriver/drive_to_point/drive_to_point.py
@@ -1,27 +1,15 @@
-import logging
-import logging.config
 import threading
 import time
 import math
 import sys
 
 from amberclient.common import amber_client
+
 from amberclient.location import location
 from amberclient.roboclaw import roboclaw
-import os
-
-from amberdriver.common import runtime
-from amberdriver.common.message_handler import MessageHandler
-from amberdriver.drive_to_point import drive_to_point_pb2
-from amberdriver.tools import config
 
 
 __author__ = 'paoolo'
-
-LOGGER_NAME = 'DriveToPointController'
-pwd = os.path.dirname(os.path.abspath(__file__))
-logging.config.fileConfig('%s/drive_to_point.ini' % pwd)
-config.add_config_ini('%s/drive_to_point.ini' % pwd)
 
 
 class DriveToPoint(object):
@@ -216,138 +204,3 @@ class DriveToPoint(object):
             _robo_angle = DriveToPoint._normalize_angle(_robo_angle)
 
         return _x, _y, _robo_angle
-
-
-class DriveToPointController(MessageHandler):
-    def __init__(self, pipe_in, pipe_out):
-        super(DriveToPointController, self).__init__(pipe_in, pipe_out)
-        self.__drive_to_point = DriveToPoint()
-
-        self.__logger = logging.getLogger(LOGGER_NAME)
-
-        runtime.add_shutdown_hook(self.terminate)
-
-    def handle_data_message(self, header, message):
-        if message.HasExtension(drive_to_point_pb2.setTargets):
-            self.__handle_set_targets(header, message)
-
-        elif message.HasExtension(drive_to_point_pb2.getNextTarget):
-            self.__handle_get_next_target(header, message)
-
-        elif message.HasExtension(drive_to_point_pb2.getNextTargets):
-            self.__handle_get_next_targets(header, message)
-
-        elif message.HasExtension(drive_to_point_pb2.getVisitedTarget):
-            self.__handle_get_visited_target(header, message)
-
-        elif message.HasExtension(drive_to_point_pb2.getVisitedTargets):
-            self.__handle_get_visited_targets(header, message)
-
-        elif message.HasExtension(drive_to_point_pb2.getConfiguration):
-            self.__handle_get_configuration(header, message)
-
-        else:
-            self.__logger.warning('No request in message')
-
-    def __handle_set_targets(self, header, message):
-        self.__logger.debug('Set targets')
-        targets = message.Extensions[drive_to_point_pb2.targets]
-        targets = zip(targets.longitudes, targets.latitudes, targets.radiuses)
-        self.__drive_to_point.set_targets(targets)
-
-    @MessageHandler.handle_and_response
-    def __handle_get_next_target(self, received_header, received_message, response_header, response_message):
-        self.__logger.debug('Get next target')
-        _next_target, _current_location = self.__drive_to_point.get_next_target_and_location()
-
-        _targets = response_message.Extensions[drive_to_point_pb2.targets]
-        _targets.longitudes.extend([_next_target[0]])
-        _targets.latitudes.extend([_next_target[1]])
-        _targets.radiuses.extend([_next_target[2]])
-
-        _location = response_message.Extensions[drive_to_point_pb2.location]
-        _location.x, _location.y, _location.p, _location.alfa, _location.timeStamp = _current_location
-
-        response_message.Extensions[drive_to_point_pb2.getNextTarget] = True
-
-        return response_header, response_message
-
-    @MessageHandler.handle_and_response
-    def __handle_get_next_targets(self, received_header, received_message, response_header, response_message):
-        self.__logger.debug('Get next targets')
-        next_targets, _current_location = self.__drive_to_point.get_next_targets_and_location()
-
-        _targets = response_message.Extensions[drive_to_point_pb2.targets]
-        _targets.longitudes.extend(map(lambda next_target: next_target[0], next_targets))
-        _targets.latitudes.extend(map(lambda next_target: next_target[1], next_targets))
-        _targets.radiuses.extend(map(lambda next_target: next_target[2], next_targets))
-
-        _location = response_message.Extensions[drive_to_point_pb2.location]
-        _location.x, _location.y, _location.p, _location.alfa, _location.timeStamp = _current_location
-
-        response_message.Extensions[drive_to_point_pb2.getNextTargets] = True
-
-        return response_header, response_message
-
-    @MessageHandler.handle_and_response
-    def __handle_get_visited_target(self, received_header, received_message, response_header, response_message):
-        self.__logger.debug('Get visited target')
-        visited_target, _current_location = self.__drive_to_point.get_visited_target_and_location()
-
-        _targets = response_message.Extensions[drive_to_point_pb2.targets]
-        _targets.longitudes.extend([visited_target[0]])
-        _targets.latitudes.extend([visited_target[1]])
-        _targets.radiuses.extend([visited_target[2]])
-
-        _location = response_message.Extensions[drive_to_point_pb2.location]
-        _location.x, _location.y, _location.p, _location.alfa, _location.timeStamp = _current_location
-
-        response_message.Extensions[drive_to_point_pb2.getVisitedTarget] = True
-
-        return response_header, response_message
-
-    @MessageHandler.handle_and_response
-    def __handle_get_visited_targets(self, received_header, received_message, response_header, response_message):
-        self.__logger.debug('Get visited targets')
-        visited_targets, _current_location = self.__drive_to_point.get_visited_targets_and_location()
-
-        _targets = response_message.Extensions[drive_to_point_pb2.targets]
-        _targets.longitudes.extend(map(lambda target: target[0], visited_targets))
-        _targets.latitudes.extend(map(lambda target: target[1], visited_targets))
-        _targets.radiuses.extend(map(lambda target: target[2], visited_targets))
-
-        _location = response_message.Extensions[drive_to_point_pb2.location]
-        _location.x, _location.y, _location.p, _location.alfa, _location.timeStamp = _current_location
-
-        response_message.Extensions[drive_to_point_pb2.getVisitedTargets] = True
-
-        return response_header, response_message
-
-    @MessageHandler.handle_and_response
-    def __handle_get_configuration(self, received_header, received_message, response_header, response_message):
-        self.__logger.debug('Get configuration')
-
-        _configuration = response_message.Extensions[drive_to_point_pb2.configuration]
-        _configuration.maxSpeed = self.__drive_to_point.MAX_SPEED
-
-        response_message.Extensions[drive_to_point_pb2.getConfiguration] = True
-
-        return response_header, response_message
-
-    def handle_subscribe_message(self, header, message):
-        self.__logger.debug('Subscribe action, nothing to do...')
-
-    def handle_unsubscribe_message(self, header, message):
-        self.__logger.debug('Unsubscribe action, nothing to do...')
-
-    def handle_client_died_message(self, client_id):
-        self.__logger.info('Client %d died, nothing to do...' % client_id)
-
-    def terminate(self):
-        self.__logger.warning('drive_to_point: terminate')
-        self.__drive_to_point.terminate()
-
-
-if __name__ == '__main__':
-    controller = DriveToPointController(sys.stdin, sys.stdout)
-    controller()
