+++ src/amberdriver/tests/hokuyo_test.py
@@ -1,0 +1,8 @@
+__author__ = 'paoolo'
+
+import unittest
+
+
+class HokuyoControllerTestCase(unittest.TestCase):
+    def setUp(self):
+        pass
